==> notebooks/01-feature-engineering.ipynb <==
# Feature Engineering Github Data

# import libraries
import pandas as pd
import numpy as np

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# read the data
#df = pd.read_csv("../data/total.csv")
df = pd.read_csv("../data/raw_data.csv")

# checking the data shape
df.shape

### One hot-encoding

# import preprocessing from sklearn
from sklearn import preprocessing
le = preprocessing.LabelEncoder()

df.language.value_counts()

# encoding these columns that are 0/1
df['private'] = le.fit_transform(df.private.values)
df['fork'] = le.fit_transform(df.fork.values)
df['has_issues'] = le.fit_transform(df.has_issues.values)
df['has_downloads'] = le.fit_transform(df.has_downloads.values)
df['has_wiki'] = le.fit_transform(df.has_wiki.values)
df['admin'] = le.fit_transform(df.admin.values)
df['push'] = le.fit_transform(df.push.values)
df['pull'] = le.fit_transform(df.pull.values)

# encoding the language categorical feature
#df = pd.get_dummies(df, columns = ["language", "label", "type"])
df = pd.get_dummies(df, columns = ["language", "type"])

df.shape

### Total Files

# create a new column called total_files that sums up the filenames column
df['total_files'] = df.apply(lambda row: row.filenames.count(',') + 1, axis=1)

### Y Creation

# use that paper to classify the commits as being defective of not
df['results'] = np.where(df['message'].str.lower().str.contains("bug|fix|error|wrong|fail|problem|patch"), 
                         1, 
                         0)

#### Delete null date column and date that is equal to 1

df = df[pd.notnull(df['date'])]
df = df.drop(df[(df.date == '1')].index)

### Order by the name and date

df = df.sort_values('name')
df = df.sort_values('date')

#### Reset the Index

df = df.reset_index(drop=True)

### Find the Weekend

import datetime
import re

def transformStringDate(strDate):
    match = re.search(r'\d{4}-\d{2}-\d{2}', strDate)
    date = datetime.datetime.strptime(match.group(), '%Y-%m-%d').date()
    return date.strftime('%A')

df['date_only'] = df['date'].apply(transformStringDate)

def weekendcheck(s):
    if s == 'Saturday' or s == 'Sunday': 
        return 1
    else:
        return 0

df['is_weekend'] = df['date_only'].apply(weekendcheck)

### Find the night

def isnight(datetimeobject):
    match = re.search('\d{2}:\d{2}:\d{2}', datetimeobject).group(0)
    first_two_letters = match[:2]
    if int(first_two_letters) > 20 or int(first_two_letters) < 6:
        return(1)
    else:
        return(0)

df['is_night'] = df['date'].apply(isnight)

### Extract the Owner

def checkOwner(fullname):
    data = fullname.split("/")
    
    if(data[0] == data[1]):
        return(0)
    else:
        return(1)

df['outside_contribution'] = df['fullname'].apply(checkOwner)
#df['outside_contribution'] = df.apply(lambda x: checkOwner(x.fullname, x.owner), axis=1)

### Check for Tests

def checkTests(filesCommitted): 
    #for names in filesCommitted:
    if (filesCommitted.find('test') != -1 or filesCommitted.find('Test') != -1): 
        return 1 

    return 0

df['tests_included'] = df['filenames'].apply(checkTests)

df.tests_included.value_counts()

### Check the Changes

from collections import Counter

def checkChanges(fileStatus): 
    list_files = fileStatus.split(",")
    occurence = Counter(list_files).most_common(1)
    tupleOccurences = occurence[0]
    if (tupleOccurences[0].find('modified') != -1): 
        return 'modified'
    elif (tupleOccurences[0].find('added') != -1): 
        return 'added'
    elif (tupleOccurences[0].find('removed') != -1): 
        return 'removed'
    elif (tupleOccurences[0].find('renamed') != -1): 
        return 'renamed'
    else:
        return 'unknown'

df['most_changes'] = df['status'].apply(checkChanges)

df.most_changes.unique()

# encoding the language categorical feature
df = pd.get_dummies(df, columns = ["most_changes"])

#df.head(5)

df.columns

# save the csv
df.to_csv('../data/total_processed.csv')
