==> notebooks/03-data-exploration.ipynb <==
# Data Exploration Github Data

#### Read the Dataframe

# import libraries
import pandas as pd
import numpy as np

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# read the data
df = pd.read_csv("../data/total_transformed.csv",
                 index_col='date', 
                 parse_dates=['date'],
                 infer_datetime_format=True)

#### Delete some columns from my dataframe

# dropping the columns that we do not need
df.drop('sha', axis=1, inplace=True)
df.drop('message', axis=1, inplace=True)
df.drop('status', axis=1, inplace=True)
df.drop('name', axis=1, inplace=True)
df.drop('fullname', axis=1, inplace=True)
df.drop('filenames', axis=1, inplace=True)
df.drop('comments', axis=1, inplace=True)
df.drop('additions', axis=1, inplace=True)
df.drop('deletions', axis=1, inplace=True)
df.drop('changes', axis=1, inplace=True)
df.drop('owner', axis=1, inplace=True)
df.drop('open_issues_count', axis=1, inplace=True)
df.drop('forks_count', axis=1, inplace=True)
df.drop('watchers_count', axis=1, inplace=True)
df.drop('date_only', axis=1, inplace=True)
df.drop('variance_commits', axis=1, inplace=True)
df.drop('results', axis=1, inplace=True)
df.drop('label', axis=1, inplace=True)

# save the csv
df.to_csv('../data/total_dry_no_info_phases.csv')

df.head(3)
