==> notebooks/04-derivate-data.ipynb <==
# Derivate Columns

#### Read the dataset

# import the libraries
import pandas as pd
import numpy as np
import sys

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# read the data
df = pd.read_csv("../data/total_dry.csv", index_col=0)

df = df.loc[:, ~df.columns.str.contains('^Unnamed')]

# convert all columns of DataFrame
df = df.dropna()

df = df.apply(pd.to_numeric)

df.head(3)

#### Derivate all columns

def derive(items):
    i = 1
    result = []
    result.append(0)

    while i < len(items):
        if items[i-1] == 0:
            result.append((items[i]-items[i-1]))
            i = i + 1
        else:
            result.append((items[i]-items[i-1])/items[i-1])
            i = i + 1

    return result

for column in df.columns:
    col_values = derive(df[column].values)
    df[column+'_derivate'] = col_values

df = df.drop("commit_derivate", axis=1)

results = df.commit.values

df = df.drop("commit", axis=1)

df['commit'] = results 

df.commit.value_counts()

df.to_csv('commits_transformed_new.csv')
